fix: Allow reading of files without suffixes & Modify entryCode get logic
